Moving AML e2e tests from agent-metric-logs to agent-metrics-logs
