+++ test/new-e2e/tests/agent-metrics-logs/kindfilelogging/kind.go
