+++ test/new-e2e/tests/agent-metrics-logs/log-agent/k8s-logs/file_tailing_cca_off_test.go
@@ -19,7 +19,7 @@ import (
 
 	"github.com/DataDog/datadog-agent/test/new-e2e/pkg/e2e"
 	"github.com/DataDog/datadog-agent/test/new-e2e/pkg/environments"
-	kindfilelogger "github.com/DataDog/datadog-agent/test/new-e2e/tests/agent-metric-logs/kindfilelogging"
+	kindfilelogger "github.com/DataDog/datadog-agent/test/new-e2e/tests/agent-metrics-logs/kindfilelogging"
 )
 
 type k8sCCAOffSuite struct {
