+++ test/new-e2e/tests/agent-metrics-logs/log-agent/linux-log/file-tailing/log-config/config.yaml
