+++ test/new-e2e/tests/agent-metrics-logs/log-agent/linux-log/journald/log-config/logger-service.sh
