+++ test/new-e2e/tests/agent-metrics-logs/log-agent/listener/testdata/agent-logs-tcp-conf.yaml
