+++ test/new-e2e/tests/agent-metrics-logs/log-agent/utils/file_tailing_utils.go
