+++ test/new-e2e/tests/agent-metrics-logs/log-agent/windows-log/file-tailing/log-config/config.yaml
